Refactor friendships_spec
1. Add friendships and friendship route to request_routes spec helper module
2. Use request routes helper methods in friendships_spec
3. Refactor all specs to include unauthenticated requests and also
complete coverage
4. Refactor friendships_controller to use application_controller json
rendering methods as well as to simplify actions

+++ app/controllers/v1/friendships_controller.rb
@@ -23,25 +23,31 @@ class V1::FriendshipsController < ApplicationController
 
   def destroy
     friendship = find_friendship
+    return unless friendship
+
     friendship.destroy
-    if friendship.confirmed
-      render json: { message: 'Friendship deleted' }, status: :accepted
-    else
-      message = friendship.active_friend == pundit_user ? 'Cancelled friend request' : 'Rejected friend request'
-      render json: { message: message }, status: :accepted
-    end
+    delete_request_message(friendship)
   end
 
   private
 
   def find_friendship
     friendship = Friendship.order_created.find_by(id: params[:id])
+
     if friendship
       authorize friendship
       return friendship
+    end
+    find_error('friendship or request')
+    nil
+  end
+
+  def delete_request_message(friendship)
+    if friendship.confirmed
+      action_success('Friendship deleted')
     else
-      render json: { message: 'Cannot find resource' }, status: 404
-      return false
+      message = friendship.active_friend == pundit_user ? 'Cancelled friend request' : 'Rejected friend request'
+      action_success(message)
     end
   end
 end
